update excel issues
