+++ app/Exports/FinalCalculationVerticalExport.php
@@ -4,6 +4,7 @@ namespace App\Exports;
 
 use Carbon\CarbonPeriod;
 use App\Models\ProjectTimeData;
+use App\Traits\ExcelHelperTrait;
 use Illuminate\Support\Collection;
 use Maatwebsite\Excel\Events\AfterSheet;
 use PhpOffice\PhpSpreadsheet\Style\Fill;
@@ -35,6 +36,8 @@ class FinalCalculationVerticalExport implements
     }
 
     use Exportable, RegistersEventListeners;
+    use ExcelHelperTrait;
+
 
     public function title(): string
     {
@@ -379,18 +382,4 @@ class FinalCalculationVerticalExport implements
         ];
     }
 
-
-    private function getCharacterAt(int $index): string
-    {
-        $letters = range('A', 'Z');
-
-        if ($index > 26) {
-            $reminder = $index % 26;
-            $quotient = (int)floor($index / 26);
-
-            return $letters[$quotient - 1] . $letters[$reminder - 1];
-        } else {
-            return $letters[$index - 1];
-        }
-    }
 }
